requirements.txt
# Web & UI
streamlit==1.31.1
Flask==3.1.3
streamlit_mic_recorder==0.0.8

# AI, Speech & Audio
TTS==0.22.0
openai-whisper==20231117
SpeechRecognition==3.14.5
librosa==0.10.0
transformers==5.2.0
spacy==3.8.11
nltk==3.9.3

# Deep Learning (Core)
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2

# Data Science
pandas==1.5.3
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.7.


python version 3.10.11 i used to made the project 
